LR4/requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
certifi==2025.10.5
click==8.3.0
decompyle3==3.9.3
dnspython==2.8.0
email-validator==2.3.0
Faker==37.11.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
litestar==2.18.0
litestar-htmx==0.5.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgspec==0.19.0
multidict==6.7.0
multipart==1.3.0
packaging==25.0
pluggy==1.6.0
polyfactory==2.22.3
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==5.0.0
pytest-xdist==3.6.1
python-dotenv==1.1.1
PyYAML==6.0.3
rich==14.2.0
rich-click==1.9.3
six==1.17.0
sniffio==1.3.1
spark-parser==1.9.0
SQLAlchemy==2.0.44
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xdis==6.1.8
